+++ trunk/Common/RaceDefinition/Race.cs
@@ -60,11 +60,11 @@ namespace Nova.Common
         // These parameters affect the production rate of each star (used in the
         // Star class Update method).
         public int FactoryBuildCost;        // defined in the Race Designer as the amount of Resourcesrequired to build one factory
-        public double ColonistsPerResource;
-        public double FactoryProduction;    // defined in the Race Designer as the amount of resources produced by 10 factories
-        public double OperableFactories;
-        public double MineProductionRate;   // defined in the Race Designer as the amount of minerals (kT) mined by every 10 mines
-        public double OperableMines;
+        public int ColonistsPerResource;
+        public int FactoryProduction;    // defined in the Race Designer as the amount of resources produced by 10 factories
+        public int OperableFactories;
+        public int MineProductionRate;   // defined in the Race Designer as the amount of minerals (kT) mined by every 10 mines
+        public int OperableMines;
 
         public double GrowthRate;
